Renamed celery to celery_tasks
